requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==5.0.2
django-bootstrap4==24.1
django-crispy-forms==2.1
pillow==10.2.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
